backend/requirements.txt
# ==================================================
# DocMind Agent - Python Dependencies
# ==================================================

# --- Core Framework ---
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0

# --- Database ---
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
psycopg2-binary==2.9.9
alembic==1.13.0

# --- Redis ---
redis[hiredis]==5.1.0

# --- Authentication ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# --- Celery ---
celery[redis]==5.4.0
flower==2.0.1

# --- Object Storage ---
minio==7.2.0
marshmallow==3.23.3

# --- Vector Database ---
pymilvus==2.4.0

# --- Elasticsearch ---
elasticsearch[async]==8.13.0

# --- Neo4j ---
neo4j==5.18.0

# --- HTTP Client ---
httpx==0.27.0
pypdf==5.4.0

# --- Logging ---
structlog==24.4.0

# --- Linting & Testing ---
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-cov==5.0.0
ruff==0.6.0
mypy==1.11.0

# --- Runtime Graph & Observability ---
langgraph==0.2.34
langgraph-checkpoint-postgres==2.0.25
psycopg[binary]==3.3.3
langfuse==2.50.0

# --- Security ---
presidio-analyzer==2.2.355
presidio-anonymizer==2.2.355
unstructured==0.15.0
unstructured-inference
pdfminer.six==20240706
pdf2image==1.17.0
pillow-heif==1.1.1
paddleocr==2.8.0
paddlepaddle==3.3.0

setuptools>=68.0.0
